friends/requirements.txt
asgiref==3.8.1
boto3==1.34.132
botocore==1.34.132
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
django-redis==5.4.0
django_neomodel==0.2.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pika==1.3.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.32.2
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
